+++ frontend/messages/en-US/common.php
@@ -84,6 +84,7 @@ return [
     'edit app logic' => '',
     'manage' => '',
     'manage developers' => '',
+    'owner' => '',
     'view' => '',
     'Other' => '@@@@',
 ];